Added a new feature to split lines by passing a character as argument
